+++ assets/javascripts/kitten/karl/pages/partner/stories.js
@@ -1,0 +1,5 @@
+import React from 'react'
+import { storiesOf } from '@storybook/react'
+import PartnerPage from './components/page'
+
+storiesOf('Pages/Partner', module).add('PartnerPage', () => <PartnerPage />)