+++ src/helpers/react/react-elements.js
@@ -5,7 +5,7 @@ var _interopRequireDefault = require("@babel/runtime/helpers/interopRequireDefau
 Object.defineProperty(exports, "__esModule", {
   value: true
 });
-exports.getReactElementsByType = void 0;
+exports.getReactElementsWithoutType = exports.getReactElementsByType = void 0;
 
 var _react = _interopRequireDefault(require("react"));
 
@@ -18,3 +18,13 @@ var getReactElementsByType = function getReactElementsByType(_ref) {
 };
 
 exports.getReactElementsByType = getReactElementsByType;
+
+var getReactElementsWithoutType = function getReactElementsWithoutType(_ref2) {
+  var children = _ref2.children,
+      type = _ref2.type;
+  return _react.default.Children.toArray(children).filter(function (child) {
+    return child.type !== type;
+  });
+};
+
+exports.getReactElementsWithoutType = getReactElementsWithoutType;